requirements.txt
altgraph==0.17.3
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
filelock==3.12.0
gunicorn==21.2.0
idna==3.4
oauthlib==3.2.2
packaging==23.1
pefile==2023.2.7
platformdirs==3.5.1
pycparser==2.21
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
pywin32-ctypes==0.2.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.23.0
whitenoise==6.5.0
